requirements.txt
flask==2.3.3
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
torch>=2.2.0
torchvision>=0.17.0
opencv-python==4.8.0.76
numpy>=1.24.3
Pillow>=9.5.0
python-dotenv==1.0.0
diffusers>=0.21.4
transformers>=4.33.2
accelerate>=0.23.0
controlnet_aux>=0.0.7
xformers>=0.0.22
scipy>=1.11.0
safetensors>=0.4.0
tqdm>=4.66.0
ftfy>=6.1.1
spacy>=3.5.0
pytorch_lightning>=2.0.0
albumentations>=1.3.0
einops>=0.6.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
imageio>=2.31.0
invisible-watermark>=0.1.0
gradio>=4.0.0